Move Env out of js-compute-builtins.cpp
