+++ c-dependencies/js-compute-runtime/builtins/env.h
@@ -1,0 +1,10 @@
+#ifndef JS_COMPUTE_RUNTIME_BUILTIN_ENV_H
+#define JS_COMPUTE_RUNTIME_BUILTIN_ENV_H
+
+#include "builtin.h"
+
+namespace Env {
+JSObject *create(JSContext *cx);
+}
+
+#endif
